session and supercollider packages
