+++ go-webrtc-server/main.go
@@ -1,22 +1,14 @@
 package main
 
 import (
-	"bufio"
-	"bytes"
 	"encoding/json"
 	"flag"
 	"fmt"
 	"log"
-	"math/rand"
-	"net"
 	"net/http"
 	"os"
 	"os/exec"
-	"path/filepath"
-	"strconv"
 	"strings"
-	"sync"
-	"time"
 
 	"github.com/gorilla/mux"
 	"github.com/hypebeast/go-osc/osc"
@@ -26,6 +18,9 @@ import (
 
 	gst "github.com/po-studio/go-webrtc-server/internal/gstreamer-src"
 	"github.com/po-studio/go-webrtc-server/internal/signal"
+
+	"github.com/po-studio/go-webrtc-server/session"
+	sc "github.com/po-studio/go-webrtc-server/supercollider"
 )
 
 type BrowserOffer struct {
@@ -33,91 +28,25 @@ type BrowserOffer struct {
 	Type string `json:"type"`
 }
 
-type AppSession struct {
-	Id                string
-	PeerConnection    *webrtc.PeerConnection
-	GStreamerPipeline *gst.Pipeline
-	SuperColliderCmd  *exec.Cmd
-	SuperColliderPort int
-	AudioSrc          *string
-}
-
-type SessionManager struct {
-	Sessions map[string]*AppSession
-	mutex    sync.Mutex
-}
-
-func (sm *SessionManager) CreateSession(id string) *AppSession {
-	sm.mutex.Lock()
-	defer sm.mutex.Unlock()
-
-	appSession := &AppSession{}
-	appSession.Id = id
-
-	audioSrcFlag := fmt.Sprintf("audio-src-%s", id)
-	audioSrcConfig := fmt.Sprintf("jackaudiosrc name=%s ! audioconvert ! audioresample", id)
-
-	appSession.AudioSrc = flag.String(audioSrcFlag, audioSrcConfig, "GStreamer audio src")
-
-	sm.Sessions[id] = appSession
-
-	return appSession
-}
-
-func (sm *SessionManager) GetSession(id string) (*AppSession, bool) {
-	sm.mutex.Lock()
-	defer sm.mutex.Unlock()
-	session, exists := sm.Sessions[id]
-	return session, exists
-}
-
-func (sm *SessionManager) DeleteSession(id string) {
-	sm.mutex.Lock()
-	defer sm.mutex.Unlock()
-	delete(sm.Sessions, id)
-}
-
-var sessionManager = SessionManager{
-	Sessions: make(map[string]*AppSession),
-}
-
-var synthDefDirectory = "/app/supercollider/synthdefs"
-
 func main() {
 	signalChannel := make(chan os.Signal, 1)
 
 	go func() {
 		<-signalChannel
 		log.Println("Received shutdown signal. Stopping processes...")
-
-		// revisit...
-		// need all appSessions to do this gracefully
-		// stopAllProcesses()
-
 		os.Exit(0)
 	}()
 
 	router := mux.NewRouter()
+
 	router.HandleFunc("/offer", handleOffer).Methods("POST")
 	router.HandleFunc("/stop", handleStop).Methods("POST")
 	router.HandleFunc("/", serveHome).Methods("GET")
+
 	router.PathPrefix("/").Handler(http.FileServer(http.Dir("./client/")))
+
 	fmt.Println("Server started at http://localhost:8080")
 	log.Fatal(http.ListenAndServe(":8080", router))
-}
-
-func findAvailableSuperColliderPort() (int, error) {
-	addr, err := net.ResolveUDPAddr("udp", "localhost:0")
-	if err != nil {
-		return 0, err
-	}
-
-	l, err := net.ListenUDP("udp", addr)
-	if err != nil {
-		return 0, err
-	}
-	defer l.Close()
-	return l.LocalAddr().(*net.UDPAddr).Port, nil
 }
 
 func serveHome(w http.ResponseWriter, r *http.Request) {
@@ -127,9 +56,8 @@ func serveHome(w http.ResponseWriter, r *http.Request) {
 func handleStop(w http.ResponseWriter, r *http.Request) {
 	fmt.Println("Stop signal received, cleaning up...")
 
-	appSession, ok := getOrCreateSession(r, w)
+	appSession, ok := session.GetOrCreateSession(r, w)
 	if !ok {
-		// If getOrCreateSession returned false, an error has already been sent to the client
 		return
 	}
 
@@ -137,40 +65,12 @@ func handleStop(w http.ResponseWriter, r *http.Request) {
 	w.WriteHeader(http.StatusOK)
 }
 
-func getSessionIDFromHeader(r *http.Request) (string, bool) {
-	sessionID := r.Header.Get("X-Session-ID") // Custom header, typically prefixed with 'X-'
-	if sessionID == "" {
-		return "", false // No session ID found in headers
-	}
-	return sessionID, true
-}
-
-func getOrCreateSession(r *http.Request, w http.ResponseWriter) (*AppSession, bool) {
-	// Extract session ID from the HTTP header
-	sessionID, ok := getSessionIDFromHeader(r)
-	if !ok {
-		log.Println("No session ID provided in the header")
-		http.Error(w, "No session ID provided", http.StatusBadRequest)
-		return nil, false
-	}
-
-	// Retrieve or create a session based on the session ID
-	appSession, exists := sessionManager.GetSession(sessionID)
-	if !exists {
-		// Optionally create a new session if one does not exist
-		appSession = sessionManager.CreateSession(sessionID)
-	}
-
-	return appSession, true
-}
-
 func handleOffer(w http.ResponseWriter, r *http.Request) {
 	flag.Parse()
 
 	var err error
-	appSession, ok := getOrCreateSession(r, w)
+	appSession, ok := session.GetOrCreateSession(r, w)
 	if !ok {
-		// If getOrCreateSession returned false, an error has already been sent to the client
 		return
 	}
 
@@ -278,11 +178,11 @@ func handleOffer(w http.ResponseWriter, r *http.Request) {
 	<-pipelineReady
 
 	log.Println("Starting SuperCollider...")
-	go startSuperCollider(appSession)
+	go sc.StartSuperCollider(appSession)
 
 	<-gatherComplete
 
-	sendPlaySynthMessage(appSession.SuperColliderPort)
+	sc.SendPlaySynthMessage(appSession.SuperColliderPort)
 
 	localDescription := appSession.PeerConnection.LocalDescription()
 	encodedLocalDesc, err := json.Marshal(localDescription)
@@ -298,179 +198,8 @@ func handleOffer(w http.ResponseWriter, r *http.Request) {
 	w.Write(encodedLocalDesc)
 }
 
-func getRandomSynthDefName() (string, error) {
-	files, err := os.ReadDir(synthDefDirectory)
-	if err != nil {
-		return "", err
-	}
-
-	var synthDefs []string
-	for _, file := range files {
-
-		// Trim the '.scsyndef' extension from the filename before adding to the list
-		if filepath.Ext(file.Name()) == ".scsyndef" {
-			baseName := strings.TrimSuffix(file.Name(), ".scsyndef")
-			synthDefs = append(synthDefs, baseName)
-		}
-	}
-
-	if len(synthDefs) == 0 {
-		return "", fmt.Errorf("no .scsyndef files found in %s", synthDefDirectory)
-	}
-
-	randTimeSeed := rand.NewSource(time.Now().UnixNano())
-	rnd := rand.New(randTimeSeed)
-
-	chosenFile := synthDefs[rnd.Intn(len(synthDefs))]
-
-	return chosenFile, nil
-}
-
-func sendPlaySynthMessage(port int) {
-	client := osc.NewClient("127.0.0.1", port)
-	msg := osc.NewMessage("/s_new")
-
-	synthDefName, err := getRandomSynthDefName()
-	if err != nil {
-		log.Printf("Could not find synthdef name: %v", err)
-		return
-	}
-
-	msg.Append(synthDefName)
-
-	msg.Append(int32(1)) // node ID
-	msg.Append(int32(0)) // action: 0 for add to head
-	msg.Append(int32(0)) // target group ID
-
-	log.Printf("Sending OSC message: %v", msg)
-	if err := client.Send(msg); err != nil {
-		log.Printf("Error sending OSC message: %v\n", err)
-	} else {
-		log.Println("OSC message sent successfully.")
-	}
-}
-
-func startSuperCollider(appSession *AppSession) {
-	scPort, err := findAvailableSuperColliderPort()
-	if err != nil {
-		log.Printf("Error finding SuperCollider port: %v", err)
-		return
-	}
-	appSession.SuperColliderPort = scPort
-
-	gstJackPorts, err := setGStreamerJackPorts(appSession)
-	if err != nil {
-		log.Printf("Error finding GStreamer-JACK ports: %v", err)
-		return
-	}
-	gstJackPortsStr := strings.Join(gstJackPorts, ",")
-
-	cmd := exec.Command(
-		"scsynth",
-		"-u", strconv.Itoa(scPort),
-		"-a", "1024",
-		"-i", "2",
-		"-o", "2",
-		"-b", "1026",
-		"-R", "0",
-		"-C", "0",
-		"-l", "1",
-	)
-	cmd.Env = append(os.Environ(),
-		"SC_JACK_DEFAULT_OUTPUTS="+gstJackPortsStr,
-		"SC_SYNTHDEF_PATH="+synthDefDirectory,
-	)
-
-	// Open a file to log scsynth output
-	logFile, err := os.OpenFile("/app/scsynth_"+appSession.Id+".log", os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
-	if err != nil {
-		log.Printf("Failed to open log file: %v", err)
-		return
-	}
-	defer logFile.Close()
-
-	cmd.Stdout = logFile
-	cmd.Stderr = logFile
-
-	// Start the command
-	if err := cmd.Start(); err != nil {
-		log.Printf("Failed to start scsynth: %v", err)
-		return
-	}
-	appSession.SuperColliderCmd = cmd
-	log.Println("scsynth command started with dynamically assigned port:", scPort)
-
-	go monitorSCSynthOutput(logFile, appSession.SuperColliderPort)
-}
-
-// TODO find a better way to set ports from jack_lsp
-func setGStreamerJackPorts(appSession *AppSession) ([]string, error) {
-	cmd := exec.Command("jack_lsp")
-	var out bytes.Buffer
-	cmd.Stdout = &out
-	err := cmd.Run()
-	if err != nil {
-		return nil, fmt.Errorf("error listing JACK ports: %w", err)
-	}
-
-	ports := strings.Split(out.String(), "\n")
-	var gstJackPorts []string
-	prefix := "webrtc-server"
-
-	log.Println("appSession.Id: ", appSession.Id)
-	for _, port := range ports {
-		if strings.HasPrefix(port, prefix) && strings.Contains(port, appSession.Id) {
-			gstJackPorts = append(gstJackPorts, port)
-		}
-	}
-
-	return gstJackPorts, nil
-}
-
-// TODO find a deterministic way to sendPlaySynthMessage when SC is ready
-func monitorSCSynthOutput(logFile *os.File, port int) {
-	scanner := bufio.NewScanner(logFile)
-	for scanner.Scan() {
-		line := scanner.Text()
-		log.Println("SCSynth Log:", line)
-		if strings.Contains(line, "SuperCollider 3 server ready.") {
-			sendPlaySynthMessage(port)
-			break
-		}
-	}
-}
-
-// func connectJackPorts(appSession *AppSession) error {
-// 	webrtcPorts, err := setGStreamerJackPorts(appSession)
-// 	if err != nil {
-// 		return fmt.Errorf("error finding JACK ports: %w", err)
-// 	}
-
-// 	var connectErrors []string
-// 	for _, webrtcPort := range webrtcPorts {
-// 		if err := connectPort("SuperCollider:out_1", webrtcPort); err != nil {
-// 			connectErrors = append(connectErrors, err.Error())
-// 		}
-// 	}
-
-// 	if len(connectErrors) > 0 {
-// 		return fmt.Errorf("failed to connect some ports: %s", strings.Join(connectErrors, "; "))
-// 	}
-
-// 	return nil
-// }
-
-// func connectPort(outputPort, inputPort string) error {
-// 	cmd := exec.Command("jack_connect", outputPort, inputPort)
-// 	if err := cmd.Run(); err != nil {
-// 		return fmt.Errorf("failed to connect %s to %s: %w", outputPort, inputPort, err)
-// 	}
-// 	return nil
-// }
-
-// STOP LOGIC /////////////////////////////////////////////////////////
-func disconnectJackPorts(appSession *AppSession) error {
-	webrtcPorts, err := setGStreamerJackPorts(appSession)
+func disconnectJackPorts(appSession *session.AppSession) error {
+	webrtcPorts, err := sc.SetGStreamerJackPorts(appSession)
 	if err != nil {
 		return fmt.Errorf("error finding JACK ports: %w", err)
 	}
@@ -496,7 +225,7 @@ func disconnectPort(outputPort, inputPort string) error {
 	return nil
 }
 
-func stopAllProcesses(appSession *AppSession) {
+func stopAllProcesses(appSession *session.AppSession) {
 	fmt.Println("Stopping all processes...")
 
 	stopSuperCollider(appSession)
@@ -524,8 +253,7 @@ func stopAllProcesses(appSession *AppSession) {
 	fmt.Println("All processes have been stopped.")
 }
 
-// Function to stop SuperCollider gracefully
-func stopSuperCollider(appSession *AppSession) error {
+func stopSuperCollider(appSession *session.AppSession) error {
 	if appSession.SuperColliderCmd == nil || appSession.SuperColliderCmd.Process == nil {
 		fmt.Println("SuperCollider is not running")
 		return nil
